refined card styling
