+++ app/(tabs)/navigation.component.js
@@ -80,7 +80,7 @@ const styles = StyleSheet.create({
       width: 4,
       height: 4,
     },
-    shadowOpacity: 1,
+    shadowOpacity: 0.6,
     shadowRadius: 8,
     elevation: 5,
   },
@@ -92,7 +92,7 @@ const styles = StyleSheet.create({
     position: "absolute",
     bottom: 24,
     width: 264,
-    height: 80,
+    height: 72,
     zIndex: 999,
     alignSelf: "center",
     overflow: "hidden",
