+++ components/atoms/filterButton.tsx
@@ -1,40 +1,37 @@
-import React from 'react';
-import { StyleSheet } from 'react-native';
-import { IndexPath, Layout, Select, SelectItem } from '@ui-kitten/components';
+import React from "react";
+import { StyleSheet } from "react-native";
+import { IndexPath, Layout, Select, SelectItem } from "@ui-kitten/components";
 
 const data = [
   {
-     "id": 0, 
-     "name": "All",
+    id: 0,
+    name: "All",
   },
   {
-     "id": 1, 
-     "name": "Canadian Pension Plan",
+    id: 1,
+    name: "Canadian Pension Plan",
   },
   {
-     "id": 2, 
-      "name": "Disability Plan",
+    id: 2,
+    name: "Disability Plan",
   },
   {
-    "id": 3, 
-    "name": "T4 Statements",
- },
+    id: 3,
+    name: "T4 Statements",
+  },
 ];
 
 export const FilterButton = ({ onSelect }) => {
   const [selectedIndex, setSelectedIndex] = React.useState(new IndexPath(0));
-  const selectedTitle = data[selectedIndex.row]?.name || '';
+  const selectedTitle = data[selectedIndex.row]?.name || "";
 
   const handleSelect = (index) => {
     setSelectedIndex(index);
-    onSelect(index); 
+    onSelect(index);
   };
 
   return (
-    <Layout
-      style={styles.container}
-      level='1'
-    >
+    <Layout style={styles.container} level="1">
       <Select
         selectedIndex={selectedIndex}
         value={selectedTitle}
@@ -51,9 +48,9 @@ export const FilterButton = ({ onSelect }) => {
 
 const styles = StyleSheet.create({
   container: {
-    backgroundColor: 'none',
-    paddingLeft: 10, paddingRight: 10, 
-    display: 'flex',
+    backgroundColor: "none",
+    paddingLeft: 10,
+    paddingRight: 10,
+    display: "flex",
   },
-
 });
