+++ components/graveyard/gradient.js
@@ -1,0 +1,25 @@
+// import React from 'react';
+// import { View, StyleSheet } from 'react-native';
+// import { LinearGradient } from 'expo-linear-gradient';
+
+// export const GradientBackground = ({ children }) => {
+//   return (
+//     <LinearGradient
+//       colors={['#4c669f', '#3b5998', '#192f6a']}
+//       style={styles.gradientBackground}
+//     >
+//       <View style={styles.container}>
+//         {children}
+//       </View>
+//     </LinearGradient>
+//   );
+// };
+
+// const styles = StyleSheet.create({
+//   gradientBackground: {
+//     flex: 1,
+//   },
+//   container: {
+//     flex: 1,
+//   },
+// });
